Refactor: add linkpool endpoint folder structure
